requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
geopy>=2.4.0
